Added Optional course selection and more Unit tests
