+++ src/main/java/com/wewritecode/miner/GoldMiner.java
@@ -569,38 +569,4 @@ public class GoldMiner {
             return false;
         }
     }
-
-
-
-    /**
-     * Formats a JSON object nicely into a file output.
-     *
-     * @param jsonObject Object to convert to string and output into file (formatted nicely).
-     * @param filename Target file name (existing or new) for the output.
-     * @return True if it was able to write to the file, false if an IOException occurred.
-     */
-    public static boolean toJsonFile(JSONObject jsonObject, File filename, boolean doFormatting) {
-        String writeString = jsonObject.toString();
-
-        // Creates a new instance of Gson with "pretty printing" enabled.
-        if (doFormatting) {
-            Gson gson = new GsonBuilder().setPrettyPrinting().create();
-            JsonParser jp = new JsonParser();
-            JsonElement je = jp.parse(writeString);
-            writeString = gson.toJson(je);
-        }
-
-        // Write to output file
-        try {
-            FileWriter file = new FileWriter(filename);
-            file.write(writeString);
-            file.flush();
-
-        } catch (IOException e) {
-            e.printStackTrace();
-            return false;
-        }
-
-        return true;
-    }
 }
