+++ src/test/java/com/wewritecode/miner/GoldMinerTest.java
@@ -6,6 +6,7 @@ import org.junit.*;
 
 import java.io.File;
 
+import static com.wewritecode.util.JSONUtils.toJsonFile;
 import static org.junit.Assert.assertTrue;
 
 /**
@@ -21,7 +22,7 @@ public class GoldMinerTest {
     // Component Under Test (CUT)
     private static GoldMiner cut;
 
-    private static final String JSON_RESOURCE_DIR = System.getProperty("user.dir")+"/src/test/resources/jsons/";
+    private static final String JSON_RESOURCE_DIR = System.getProperty("user.dir")+"/src/test/resources/jsons/miner";
 
     @Before
     public void setup() {
@@ -40,7 +41,7 @@ public class GoldMinerTest {
         File subjectFile = new File(JSON_RESOURCE_DIR + "subjects.json");
 
         assertTrue("Expected to be able to output subject list to .json file",
-                GoldMiner.toJsonFile(actual, subjectFile, true));
+                toJsonFile(actual, subjectFile, true));
     }
 
     @Test
@@ -50,7 +51,7 @@ public class GoldMinerTest {
         actual.put("quarters", quarters);
         File subjectFile = new File(JSON_RESOURCE_DIR + "quarters.json");
         assertTrue("Expected to be able to output quarter list to .json file",
-                GoldMiner.toJsonFile(actual, subjectFile, true));
+                toJsonFile(actual, subjectFile, true));
     }
 
     @Test
@@ -59,7 +60,7 @@ public class GoldMinerTest {
         File courseLevelFile = new File(JSON_RESOURCE_DIR + "courseLevels.json");
 
         assertTrue("Expected to be able to output course levels to .json file",
-                GoldMiner.toJsonFile(courseLevels, courseLevelFile, true));
+                toJsonFile(courseLevels, courseLevelFile, true));
     }
 
     @Test
@@ -69,7 +70,7 @@ public class GoldMinerTest {
 
         File quarterFile = new File(JSON_RESOURCE_DIR + "spring2019.json");
 
-        assertTrue(GoldMiner.toJsonFile(allSubjects, quarterFile, true));
+        assertTrue(toJsonFile(allSubjects, quarterFile, true));
     }
 
     @Test
@@ -83,7 +84,7 @@ public class GoldMinerTest {
             String quarterStr = season + " " + year;
             JSONObject quarter = cut.getAllData(quarterStr);
             File quarterFile = new File(JSON_RESOURCE_DIR + season.toLowerCase() + year + ".json");
-            assertTrue(GoldMiner.toJsonFile(quarter, quarterFile, true));
+            assertTrue(toJsonFile(quarter, quarterFile, true));
         }
     }
 
@@ -98,7 +99,7 @@ public class GoldMinerTest {
             String quarterStr = season + " " + year;
             JSONObject quarter = cut.getAllData(quarterStr);
             File quarterFile = new File(JSON_RESOURCE_DIR + season.toLowerCase() + year + ".json");
-            assertTrue(GoldMiner.toJsonFile(quarter, quarterFile, true));
+            assertTrue(toJsonFile(quarter, quarterFile, true));
         }
     }
 }
