+++ src/test/resources/jsons/miner/courseLevels.json
