+++ src/test/resources/jsons/miner/quarters.json
