+++ src/test/resources/jsons/miner/spring2019.json
