+++ src/test/resources/jsons/miner/subjects.json
