+++ src/test/resources/jsons/miner/summer2019.json
