+++ src/test/resources/jsons/miner/winter2019_old.json
