+++ src/test/resources/jsons/pan/scheduleGenTest_1.json
@@ -1,0 +1,180 @@
+{
+  "mandatory": [
+    {
+      "courseID": "CS 64",
+      "lectures": [
+        {
+          "days": [
+            "M",
+            "W"
+          ],
+          "time": {
+            "start": {
+              "hour": 12,
+              "minute": 30
+            },
+            "end": {
+              "hour": 13,
+              "minute": 45
+            }
+          },
+          "sections": [
+            {
+              "days": [
+                "F"
+              ],
+              "time": {
+                "start": {
+                  "hour": 11,
+                  "minute": 0
+                },
+                "end": {
+                  "hour": 11,
+                  "minute": 50
+                }
+              }
+            },
+            {
+              "days": [
+                "F"
+              ],
+              "time": {
+                "start": {
+                  "hour": 12,
+                  "minute": 0
+                },
+                "end": {
+                  "hour": 12,
+                  "minute": 50
+                }
+              }
+            },
+            {
+              "days": [
+                "F"
+              ],
+              "time": {
+                "start": {
+                  "hour": 13,
+                  "minute": 0
+                },
+                "end": {
+                  "hour": 13,
+                  "minute": 50
+                }
+              }
+            }
+          ]
+        }
+      ]
+    },
+    {
+      "courseID": "CS 130A",
+      "lectures": [
+        {
+          "days": [
+            "M",
+            "W"
+          ],
+          "time": {
+            "start": {
+              "hour": 11,
+              "minute": 0
+            },
+            "end": {
+              "hour": 12,
+              "minute": 15
+            }
+          },
+          "sections": [
+            {
+              "days": [
+                "F"
+              ],
+              "time": {
+                "start": {
+                  "hour": 9,
+                  "minute": 0
+                },
+                "end": {
+                  "hour": 9,
+                  "minute": 50
+                }
+              }
+            },
+            {
+              "days": [
+                "F"
+              ],
+              "time": {
+                "start": {
+                  "hour": 10,
+                  "minute": 0
+                },
+                "end": {
+                  "hour": 10,
+                  "minute": 50
+                }
+              }
+            }
+          ]
+        }
+      ]
+    },
+    {
+      "courseID": "CS 138",
+      "lectures": [
+        {
+          "days": [
+            "M",
+            "W"
+          ],
+          "time": {
+            "start": {
+              "hour": 9,
+              "minute": 30
+            },
+            "end": {
+              "hour": 10,
+              "minute": 45
+            }
+          },
+          "sections": [
+            {
+              "days": [
+                "F"
+              ],
+              "time": {
+                "start": {
+                  "hour": 11,
+                  "minute": 0
+                },
+                "end": {
+                  "hour": 11,
+                  "minute": 50
+                }
+              }
+            },
+            {
+              "days": [
+                "F"
+              ],
+              "time": {
+                "start": {
+                  "hour": 12,
+                  "minute": 0
+                },
+                "end": {
+                  "hour": 12,
+                  "minute": 50
+                }
+              }
+            }
+          ]
+        }
+      ]
+    }
+  ],
+  "optional": [],
+  "filters": []
+}
